Restructuring of tests folder
