+++ frontend/src/tests/components/eMIB/HowTo.test.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { shallow } from "enzyme";
-import HowTo from "../components/eMIB/HowTo";
-import { LANGUAGES } from "../components/commons/Translation";
-import LOCALIZE from "../text_resources";
+import HowTo from "../../../components/eMIB/HowTo";
+import { LANGUAGES } from "../../../components/commons/Translation";
+import LOCALIZE from "../../../text_resources";
 
 it("renders HowTo Page Title in English", () => {
   LOCALIZE.setLanguage(LANGUAGES.english);
