+++ frontend/src/tests/components/eMIB/Tabs.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow } from "enzyme";
-import Tab, { SELECTED, UNSELECTED } from "../components/eMIB/Tab";
+import Tab, { SELECTED, UNSELECTED } from "../../../components/eMIB/Tab";
 
 it("renders selected tab", () => {
   const wrapper = shallow(<Tab tabName="Tab1" selected={SELECTED} />);
