+++ frontend/src/tests/components/status/SystemCheck.test.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { Label } from "react-bootstrap";
-import SystemCheck from "./SystemCheck";
-import LOCALIZE from "../../text_resources";
+import SystemCheck from "../../../components/status/SystemCheck";
+import LOCALIZE from "../../../text_resources";
 
 const failingMessage = <Label>{LOCALIZE.commons.failStatus}</Label>;
 const passingMessage = <Label>{LOCALIZE.commons.passStatus}</Label>;
